scmutationexec: absorb descriptorutils package
This refactoring commit moves everything in the descriptorutils package
to the scmutationexec package. It also removes unused helper functions
from the latter package.

Release note: None
