+++ pkg/sql/schemachanger/scexec/scmutationexec/helpers.go
@@ -11,106 +11,96 @@
 package scmutationexec
 
 import (
-	"context"
-
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/tabledesc"
-	"github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scexec/descriptorutils"
 	"github.com/cockroachdb/errors"
 )
 
 func mutationStateChange(
-	ctx context.Context,
-	table *tabledesc.Mutable,
-	f descriptorutils.MutationSelector,
-	exp, next descpb.DescriptorMutation_State,
+	tbl *tabledesc.Mutable, f MutationSelector, exp, next descpb.DescriptorMutation_State,
 ) error {
-	mut, err := descriptorutils.FindMutation(table, f)
+	mut, err := FindMutation(tbl, f)
 	if err != nil {
 		return err
 	}
-	m := &table.TableDesc().Mutations[mut.MutationOrdinal()]
+	m := &tbl.TableDesc().Mutations[mut.MutationOrdinal()]
 	if m.State != exp {
 		return errors.AssertionFailedf("update mutation for %d from %v to %v: unexpected state: %v",
-			table.GetID(), exp, m.State, table)
+			tbl.GetID(), exp, m.State, tbl)
 	}
 	m.State = next
 	return nil
 }
 
 func removeMutation(
-	ctx context.Context,
-	table *tabledesc.Mutable,
-	f descriptorutils.MutationSelector,
-	exp descpb.DescriptorMutation_State,
+	tbl *tabledesc.Mutable, f MutationSelector, exp descpb.DescriptorMutation_State,
 ) (descpb.DescriptorMutation, error) {
-	mut, err := descriptorutils.FindMutation(table, f)
+	mut, err := FindMutation(tbl, f)
 	if err != nil {
 		return descpb.DescriptorMutation{}, err
 	}
 	foundIdx := mut.MutationOrdinal()
-	cpy := table.Mutations[foundIdx]
+	cpy := tbl.Mutations[foundIdx]
 	if cpy.State != exp {
 		return descpb.DescriptorMutation{}, errors.AssertionFailedf(
 			"remove mutation from %d: unexpected state: got %v, expected %v: %v",
-			table.GetID(), cpy.State, exp, table,
+			tbl.GetID(), cpy.State, exp, tbl,
 		)
 	}
-	table.Mutations = append(table.Mutations[:foundIdx], table.Mutations[foundIdx+1:]...)
+	tbl.Mutations = append(tbl.Mutations[:foundIdx], tbl.Mutations[foundIdx+1:]...)
 	return cpy, nil
 }
 
-// findFamilyOrdinalForColumnID finds a family which contains the needle column
-// id and returns its index in the families slice.
-func findFamilyOrdinalForColumnID(
-	table catalog.TableDescriptor, needle descpb.ColumnID,
-) (int, error) {
-	families := table.GetFamilies()
-	for i := range families {
-		for _, colID := range families[i].ColumnIDs {
-			if colID == needle {
-				return i, nil
-			}
-		}
-	}
-	return -1, errors.Errorf("failed to find column family for column %d in table %d: %v",
-		needle, table.GetID(), table)
-}
-
-// Suppress the linter.
-var _ = findFamilyOrdinalForColumnID
-
-func removeColumnFromFamily(table *tabledesc.Mutable, colID descpb.ColumnID) error {
-	famIdx, err := findFamilyOrdinalForColumnID(table, colID)
-	if err != nil {
-		return errors.WithAssertionFailure(err)
-	}
-	f := &table.Families[famIdx]
-	for i, id := range f.ColumnIDs {
-		if id == colID {
-			f.ColumnIDs = append(f.ColumnIDs[:i], f.ColumnIDs[i+1:]...)
-			f.ColumnNames = append(f.ColumnNames[:i], f.ColumnNames[i+1:]...)
-			break
-		}
-	}
-	if len(f.ColumnIDs) == 0 {
-		table.Families = append(table.Families[:famIdx], table.Families[famIdx+1:]...)
-	}
-	return nil
-}
-
-// Suppress the linter.
-var _ = removeColumnFromFamily
-
-func columnNamesFromIDs(table *tabledesc.Mutable, columnIDs descpb.ColumnIDs) ([]string, error) {
+func columnNamesFromIDs(tbl *tabledesc.Mutable, columnIDs descpb.ColumnIDs) ([]string, error) {
 	storeColNames := make([]string, 0, len(columnIDs))
 	for _, colID := range columnIDs {
-		column, err := table.FindColumnWithID(colID)
+		column, err := tbl.FindColumnWithID(colID)
 		if err != nil {
 			return nil, err
 		}
 		storeColNames = append(storeColNames, column.GetName())
 	}
 	return storeColNames, nil
+}
+
+// MutationSelector defines a predicate on a catalog.Mutation with no
+// side-effects.
+type MutationSelector func(mutation catalog.Mutation) (matches bool)
+
+// FindMutation returns the first mutation in table for which the selector
+// returns true.
+// Such a mutation is expected to exist, if none are found, an internal error
+// is returned.
+func FindMutation(
+	tbl catalog.TableDescriptor, selector MutationSelector,
+) (catalog.Mutation, error) {
+	for _, mut := range tbl.AllMutations() {
+		if selector(mut) {
+			return mut, nil
+		}
+	}
+	return nil, errors.AssertionFailedf("matching mutation not found in table %d", tbl.GetID())
+}
+
+// MakeIndexIDMutationSelector returns a MutationSelector which matches an
+// index mutation with the correct ID.
+func MakeIndexIDMutationSelector(indexID descpb.IndexID) MutationSelector {
+	return func(mut catalog.Mutation) bool {
+		if mut.AsIndex() == nil {
+			return false
+		}
+		return mut.AsIndex().GetID() == indexID
+	}
+}
+
+// MakeColumnIDMutationSelector returns a MutationSelector which matches a
+// column mutation with the correct ID.
+func MakeColumnIDMutationSelector(columnID descpb.ColumnID) MutationSelector {
+	return func(mut catalog.Mutation) bool {
+		if mut.AsColumn() == nil {
+			return false
+		}
+		return mut.AsColumn().GetID() == columnID
+	}
 }
